+++ src/models/simulation.jl
@@ -77,7 +77,7 @@ function initialize!(
     simulation.model.output.active = output                     # enable/disable output
 
     # SCALING: we use vorticity*radius, divergence*radius in the dynamical core
-    scale!(progn, diagn, model.spectral_grid.radius)
+    scale!(progn, diagn, model.planet.radius)
 
     # OUTPUT INITIALISATION AND STORING INITIAL CONDITIONS + FEEDBACK
     # propagate spectral state to grid variables for initial condition output